🔒 chore(.env): add .env file with DB and SECRET environment variables 📦 chore(go.mod): update dependencies to latest versions The .env file was added to the project with the DB and SECRET environment variables to store sensitive information securely. The go.mod file was updated to include the latest versions of the project dependencies for improved stability and security.
📦 chore(go.sum): add new dependencies and update existing ones
The commit adds new dependencies and updates existing ones in the go.sum file to ensure that the project has the necessary dependencies for compilation and execution.

🚀 feat(loadEnvVariables.go): add function to load environment variables from .env file
🚀 feat(main.go): initialize environment variables and connect to database on app start
🚀 feat(product.go): create Product model with Id, NamaProduct, and Deskripsi fields
🚀 feat(setup.go): add ConnectDatabase function to establish connection to MySQL database
The changes introduce functionality to load environment variables from a .env file, initialize environment variables and connect to a database on app start, define a Product model with specific fields, and set up a connection to a MySQL database for the application.