🔧 fix(productcontroller): improve error handling and add support for file upload ✨ feat(productcontroller): enhance product creation process by adding support for file upload for product photo and improving error handling
The changes were made to improve the error handling in the product creation process and to add support for file upload for the product photo. This enhancement ensures better user experience and data integrity in the application.

🔧 fix(product.go): change Price, Stock, and DailyQuota data types to float64 for more precise calculations
The Price, Stock, and DailyQuota fields in the Product struct have been changed from float32 to float64 data types. This change was made to ensure more precise calculations and avoid potential loss of precision when dealing with decimal values in the application. Additionally, a new struct FileUpload with a FileHeader field of type multipart.FileHeader has been added to handle file uploads in the application.