🔧 fix(hamperscontroller.go): add missing newline at the end of the file 🚀 feat(hamperscontroller.go): implement CRUD operations for hampers and details The changes include adding CRUD operations for hampers and details in the hampers controller file. This allows for creating, updating, deleting, and retrieving hampers and their details from the database, enhancing the functionality of the application.
🔧 refactor(productcontroller): remove unused 'productRewardPoin' variable in Create and Update functions
📸 feat(images): add multiple image files for products
The 'productRewardPoin' variable was removed as it was unused in both the Create and Update functions, simplifying the code and improving readability. Additionally, multiple image files were added to enhance the product listing with visual content.

📦 add(images): add multiple image files in the images directory
🚀 feat(main.go): add new routes for hampers in the main application file
The changes involve adding multiple image files to the images directory. This allows for the inclusion of various image assets in the project. Additionally, new routes for hampers have been added in the main application file to handle different functionalities related to hampers.

🔨 refactor(models): change field names to adhere to naming conventions
The field names 'jumlah' in DetailHampers and 'RewardPoin' in Product have been changed to 'Jumlah' and 'reward_poin' respectively to follow naming conventions and improve code readability.